// client/safecycle-client/src/app/core/model/amenity.model.ts
export interface AmenityModel {
  type: string,
  id: number,
  lat: number,
  lon: number,
  tags: any
}


export enum AmenityEnum {
  BAR = "bar",
  BIERGARTEN = "biergarten",
  CAFE = "cafe",
  FAST_FOOD = "fast_food",
  FOOD_COURT = "food_court",
  ICE_CREAM = "ice_cream",
  PUB = "pub",
  RESTAURANT = "restaurant",
  COLLEGE = "college",
  DRIVING_SCHOOL = "driving_school",
  KINDERGARTEN = "kindergarten",
  LANGUAGE_SCHOOL = "language_school",
  LIBRARY = "library",
  TOY_LIBRARY = "toy_library",
  MUSIC_SCHOOL = "music_school",
  SCHOOL = "school",
  UNIVERSITY = "university",
  BICYCLE_PARKING = "bicycle_parking",
  BICYCLE_REPAIR_STATION = "bicycle_repair_station",
  BICYCLE_RENTAL = "bicycle_rental",
  BOAT_RENTAL = "boat_rental",
  BOAT_SHARING = "boat_sharing",
  BUS_STATION = "bus_station",
  CAR_RENTAL = "car_rental",
  CAR_SHARING = "car_sharing",
  CAR_WASH = "car_wash",
  VEHICLE_INSPECTION = "vehicle_inspection",
  CHARGING_STATION = "charging_station",
  FERRY_TERMINAL = "ferry_terminal",
  FUEL = "fuel",
  GRIT_BIN = "grit_bin",
  MOTORCYCLE_PARKING = "motorcycle_parking",
  PARKING = "parking",
  PARKING_ENTRANCE = "parking_entrance",
  PARKING_SPACE = "parking_space",
  TAXI = "taxi",
  ATM = "atm",
  BANK = "bank",
  BUREAU_DE_CHANGE = "bureau_de_change",
  BABY_HATCH = "baby_hatch",
  CLINIC = "clinic",
  DENTIST = "dentist",
  DOCTORS = "doctors",
  HOSPITAL = "hospital",
  NURSING_HOME = "nursing_home",
  PHARMACY = "pharmacy",
  SOCIAL_FACILITY = "social_facility",
  VETERINARY = "veterinary",
  ARTS_CENTRE = "arts_centre",
  BROTHEL = "brothel",
  CASINO = "casino",
  CINEMA = "cinema",
  COMMUNITY_CENTRE = "community_centre",
  CONFERENCE_CENTRE = "conference_centre",
  EVENTS_VENUE = "events_venue",
  FOUNTAIN = "fountain",
  GAMBLING = "gambling",
  LOVE_HOTEL = "love_hotel",
  NIGHTCLUB = "nightclub",
  PLANETARIUM = "planetarium",
  PUBLIC_BOOKCASE = "public_bookcase",
  SOCIAL_CENTRE = "social_centre",
  STRIPCLUB = "stripclub",
  STUDIO = "studio",
  SWINGERCLUB = "swingerclub",
  THEATRE = "theatre",
  COURTHOUSE = "courthouse",
  EMBASSY = "embassy",
  FIRE_STATION = "fire_station",
  POLICE = "police",
  POST_BOX = "post_box",
  POST_DEPOT = "post_depot",
  POST_OFFICE = "post_office",
  PRISON = "prison",
  RANGER_STATION = "ranger_station",
  TOWNHALL = "townhall",
  BBQ = "bbq",
  BENCH = "bench",
  DOG_TOILET = "dog_toilet",
  DRINKING_WATER = "drinking_water",
  GIVE_BOX = "give_box",
  PARCEL_LOCKER = "parcel_locker",
  SHELTER = "shelter",
  SHOWER = "shower",
  TELEPHONE = "telephone",
  TOILETS = "toilets",
  WATER_POINT = "water_point",
  WATERING_PLACE = "watering_place",
  SANITARY_DUMP_STATION = "sanitary_dump_station",
  RECYCLING = "recycling",
  WASTE_BASKET = "waste_basket",
  WASTE_DISPOSAL = "waste_disposal",
  WASTE_TRANSFER_STATION = "waste_transfer_station",
  ANIMAL_BOARDING = "animal_boarding",
  ANIMAL_BREEDING = "animal_breeding",
  ANIMAL_SHELTER = "animal_shelter",
  BAKING_OVEN = "baking_oven",
  CHILDCARE = "childcare",
  CLOCK = "clock",
  CREMATORIUM = "crematorium",
  DIVE_CENTRE = "dive_centre",
  FUNERAL_HALL = "funeral_hall",
  GRAVE_YARD = "grave_yard",
  GYM = "gym",
  HUNTING_STAND = "hunting_stand",
  INTERNET_CAFE = "internet_cafe",
  KITCHEN = "kitchen",
  KNEIPP_WATER_CURE = "kneipp_water_cure",
  LOUNGER = "lounger",
  MARKETPLACE = "marketplace",
  MONASTERY = "monastery",
  PHOTO_BOOTH = "photo_booth",
  PLACE_OF_MOURNING = "place_of_mourning",
  PLACE_OF_WORSHIP = "place_of_worship",
  PUBLIC_BATH = "public_bath",
  PUBLIC_BUILDING = "public_building",
  REFUGEE_SITE = "refugee_site",
  VENDING_MACHINE = "vending_machine",
  ALPINE_HUT = "alpine_hut",
  APARTMENT = "apartment",
  CHALET = "chalet",
  AQUARIUM = "aquarium",
  ARTWORK = "artwork",
  ATTRACTION = "attraction",
  CAMP_PITCH = "camp_pitch",
  CARAVAN_SITE = "caravan_site",
  CAMP_SITE = "camp_site",
  GALLERY = "gallery",
  MUSEUM = "museum",
  GUEST_HOUSE = "guest_house",
  HOSTEL = "hostel",
  HOTEL = "hotel",
  INFORMATION = "information",
  MOTEL = "motel",
  PICNIC_SITE = "picnic_site",
  THEME_PARK = "theme_park",
  VIEWPOINT = "viewpoint",
  WILDERNESS_HUT = "wilderness_hut",
  ZOO = "zoo",
}